board:2041 board:1992 the device decides which clusters draw, and the frame halves
THE SUBJECT PASS IS INDIRECT NOW. A compute stage sweeps one threadgroup per cluster, rejects it
against the six planes the view matrix already carries, compacts what survives into an index run
and writes the draw's own count with an atomic add. The raster pass records
`SDL_DrawGPUIndexedPrimitivesIndirect` for every batch the culler decided, and the direct draw for
the batches it could not -- a blended surface the cooker does not cut, or an instanced one whose
clusters would have to be culled per instance.

    place            before        after     places 8 of 8 green, 97.9 s against 171.9
    Shibuya       p50 59.53    p50 31.19
    Central Park  p50 28.82    p50 18.64
    Venice        p50 10.53    p50  6.36
    Heidelberg    p50  9.59    p50  6.47
    The Alps      p50  6.05    p50  5.77
    Rothenburg    p50  4.22    p50  3.21

Four of the six now hold 60 fps. Shibuya does not and the next item says what is left.

THE STAGE SITS BEFORE THE SHADOW AND THE SKY rather than in front of the draw, because a cull
immediately ahead of the pass that consumes it is a hard wait; here it shares the compute pass the
atmosphere already opens and the raster work in between covers it.

AND FOUR THINGS WERE MEASURED WRONG BEFORE THEY WERE MEASURED RIGHT, each one recorded because the
wrong version is the one that looks reasonable:

  `ResourceKind` GAINED A `Buffer` ARM and should not have. That conflated where a resource comes
  from with what its element is, and the door said so in one run: `irradiance` is a compute write
  with neither a texel format nor a stride, so a check asking every compute write for one refused a
  plan that had always been correct. `TexelFormat::Table` is the arm that was missing.

  THE INDEX RUN EXISTED SIX TIMES. 113 MB each on Shibuya: the subject's own, the cooked order
  beside it, the same run repacked in batch order, and all three again on the device. The cooker
  reorders in place now and the cluster's range travels in its JOB -- one `uint4` of cluster,
  batch, first index and count, already in the packed run's numbering -- so the kernel copies out
  of the buffer the direct path draws from and the two extra runs are gone, along with the staging
  that carried them.

  THE KERNEL GAVE A CLUSTER TO ONE THREAD and had it copy 384 indices in a loop. Heidelberg's 7650
  clusters became 7650 busy lanes on a machine built for the opposite: the frame went from 9.59 ms
  to 68.21. The copy is 23 MB a frame there, a quarter of a millisecond at this device's bandwidth
  -- all the rest was the shape of the kernel. One threadgroup per cluster decides once and every
  lane copies a strided slice.

  AND THE ONE THAT COST THE MOST TO FIND. With the cull disabled, the draw direct and the reorder
  off, Heidelberg still read 86.98 ms against the baseline's 9.60 -- the plan hash identical, the
  geometry identical, every stage's CPU time zero. `SetMesh` ended in an `OpenStaging` that widened
  the DEFERRED ring to the sum of every stream just handed over, hundreds of megabytes on a city.
  That ring is not what a rebuild uses; it is what a FRAME uses, and the first map of a frame
  CYCLES it. The cull's twenty-byte argument reset was renaming three hundred megabytes per frame
  to get written. `Cross` already widens the ring when a hand does not fit, so the guess ahead of
  the demand could only ever be wrong in the expensive direction. It is gone, and that alone is
  most of what the other five places gained.

Every one of those was found by bisecting against a BASELINE BUILD of the previous commit run on
the same machine in the same minute -- 9.60 ms measured twice, forty minutes apart, is what made
the 86.98 unarguable.

THE COOKER'S RECORD AND THE DEVICE'S RUN ARE DIFFERENT SHAPES, finally. `DagCluster` stays whole
for a cooker that walks clusters one at a time; the device gets spheres alone at a 16-byte stride,
because the sweep reads EVERY sphere and nothing else about what it rejects. Shibuya's 73 706
clusters cost that sweep 1.2 MB rather than 3.8.

AND KHRONOS DOES NOT DECIDE A LAYOUT. The in-place reorder changes which of two coincident
triangles arrives first, and NormalTangentTest has a cell that moves past its own bound because of
it. A vendor corpus states an ANSWER; it does not state a memory layout, and the importer is what
changes if that case needs it.

+++ src/base/spatial/ClusterCook.cpp
@@ -268,7 +268,7 @@ Cooked CookDag(std::span<const float> positionsM, std::span<const uint32_t> indi
     for (const DagCluster &one : above.Clusters) {
       DagCluster carried = one;
       carried.First = rebase + one.First;
-      carried.Level = (uint8_t)level;
+      carried.Level = (uint32_t)level;
       carried.SelfErr = (float)worst;
       out.Clusters.push_back(carried);
     }
